requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.2
asyncer==0.0.2
attrs==23.1.0
auth0-python==4.3.0
backoff==2.2.1
bidict==0.22.1
bleach==6.0.0
bokeh==3.2.0
cassandra-driver==3.28.0
cassio==0.0.7
certifi==2023.5.7
cffi==1.15.1
chainlit==0.5.1
charset-normalizer==3.2.0
chromadb==0.3.26
click==8.1.4
clickhouse-connect==0.6.6
coloredlogs==15.0.1
contourpy==1.1.0
cryptography==41.0.1
dataclasses-json==0.5.9
Deprecated==1.2.14
duckdb==0.8.1
exceptiongroup==1.1.2
fastapi==0.97.0
fastapi-socketio==0.0.10
filetype==1.2.0
flatbuffers==23.5.26
frozenlist==1.3.3
geomet==0.2.1.post1
googleapis-common-protos==1.59.1
grpcio==1.56.0
h11==0.14.0
hnswlib==0.7.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
humanfriendly==10.0
idna==3.4
importlib-metadata==6.0.1
Jinja2==3.1.2
langchain==0.0.228
langchainplus-sdk==0.0.20
linkify-it-py==2.0.2
lz4==4.3.2
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
mdit-py-plugins==0.4.0
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
nodeenv==1.8.0
numexpr==2.8.4
numpy==1.25.0
onnxruntime==1.15.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
opentelemetry-api==1.18.0
opentelemetry-exporter-otlp==1.18.0
opentelemetry-exporter-otlp-proto-common==1.18.0
opentelemetry-exporter-otlp-proto-grpc==1.18.0
opentelemetry-exporter-otlp-proto-http==1.18.0
opentelemetry-instrumentation==0.39b0
opentelemetry-proto==1.18.0
opentelemetry-sdk==1.18.0
opentelemetry-semantic-conventions==0.39b0
overrides==7.3.1
packaging==23.1
pandas==2.0.3
panel==1.2.0
param==1.13.0
Pillow==10.0.0
posthog==3.0.1
prisma==0.9.1
protobuf==4.23.4
pulsar-client==3.2.0
pycparser==2.21
pydantic==1.10.11
PyJWT==2.7.0
pypdf==3.12.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.5.1
python-graphql-client==0.4.3
python-socketio==5.8.0
pytz==2023.3
pyviz-comms==2.3.2
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.18
starlette==0.27.0
sympy==1.12
syncer==2.0.3
tenacity==8.2.2
tiktoken==0.4.0
tokenizers==0.13.3
tomli==2.0.1
tomlkit==0.11.8
tornado==6.3.2
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
uc-micro-py==1.0.2
uptrace==1.18.0
urllib3==2.0.3
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
webencodings==0.5.1
websockets==11.0.3
wrapt==1.15.0
xyzservices==2023.5.0
yarl==1.9.2
zipp==3.16.0
zstandard==0.21.0
